Mostly a cleanup release.  Cleanup old code, improve documentation, improve GCC-cleanness, etc.
Disable the crandom output buffer so that it won't return duplicate
data across fork().  I should still stir in more entropy into the
buffer at least when RDRAND is available, but this should prevent
disasters for now.

The Elligator code in the current version is incompatible with past
versions due to a minor tweak.  It wasn't being called by any of
the API functions, though.

Removing "magic" constants and type names.  So for example p448_t
is now field_t (though maybe it should really be felem_t?).  This
should enable other curves with the Goldilocks code in the not-too-
distant future.

Added CRANDOM_MIGHT_IS_MUST so that you don't have to -D a bunch of
things on the command line.

You can `make bat` to make an eBAT which probably doesn't work.

I haven't implemented the improved nonce generation from the
curves@moderncrypto.org thread yet.