Replace all boxed '*_t[1]' typedefs with plain structs
This library (i.e. libdecaf) originally used a "neat trick" in C to
define "unified" struct and pointer types which work in either context
by always referencing structs as 1-element arrays. This eliminated the
need to add `&` to all operands in the point and field arithmetic code,
which was visually a lot less noisy.

Some more discussion of this approach here:

https://twitter.com/syncomo/status/1029366966924890113

The main drawback of this approach is trying to build external bindings
to it. Though array and pointer types are the same in C, making things
convenient there, in the generated bindings for other languages, such as
Rust's bindgen, this approach adds a superfluous array around
everything.

Worse, since Rust bindgen can't reason about the interior mutability of
arrays, it breaks const correctness in the generated binding, since Rust
assumes all of the array interiors are mutable.

This change switches to doing the more normal/"ugly" thing and uses
ordinary structs and pointers, with the typedefs all aliased to the
structs. It adds annoying `&`s all over the arithmetic, but the
generated bindings are simple and clean, without superfluous `[1]`s.

NOTE: Lots of find/replace and hand edits, and still no tests! It's
entirely possible something was botched in this process. (The next step
after this will definitely be tests but I didn't want to write them
until the binding is const-correct)